+++ genericwave/Weapons/Weapon.h
@@ -25,13 +25,13 @@ public:
                 // Reload if possible
                 reload();
             }
-            double angle = weapon.getRotation();
-            double xdiff = cos(angle*pi/180.0) * weaponWidth;
-            double ydiff = sin(angle*pi/180.0) * weaponHeight;
+            long double angle = weapon.getRotation();
+            long double xdiff = cos(angle*pi/180.0) * weaponWidth;
+            long double ydiff = sin(angle*pi/180.0) * weaponHeight;
 
             // Position of pistol nozzle, i.e. where the bullet will be fired.
-            double bulletx = weapon.getPosition().x + xdiff;
-            double bullety = weapon.getPosition().y + ydiff;
+            long double bulletx = weapon.getPosition().x + xdiff;
+            long double bullety = weapon.getPosition().y + ydiff;
 
             //std::cout << "xdiff: " << xdiff <<std::endl;
             //std::cout << "ydiff: " << ydiff << std::endl;
